Project Archive, New Resume, Force Server Page Load
Add a project archive with the old project htmldocs I have. Might as
well, I guess.

Update my resume's alignments. It was a bit jank and bothering me.

Force the page to fetch from the server every time. Me and every other
person on the planet (probably) was having issues with caching. Boo hoo.
It's like, a megabyte. Whoop-de-fucking-do. I'll remove it if it's too
slow.